08_id_card_photo_validation/requirements.txt
azure-cognitiveservices-vision-face>=0.6.0
azure-cognitiveservices-vision-computervision>=0.9.0
azure-identity>=1.16.0
msrest>=0.7.1
Pillow>=10.3.0
python-dotenv>=1.0.0
requests>=2.32.0
pandas>=2.2.0
